requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
confluent-kafka==2.6.1
fastapi==0.115.6
h11==0.14.0
idna==3.10
mysql-connector-python==9.1.0
pydantic==2.10.3
pydantic_core==2.27.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
